Fix comment: separate MultiSet.cs file; add remark for not removing entry with 0 item count.
