+++ src/libraries/System.Collections.Immutable/src/System/Collections/Immutable/ImmutableArray_1.cs
@@ -1384,7 +1384,7 @@ namespace System.Collections.Immutable
 
         private ICollection<int>? FindOrderedIndicesToRemoveByMultiSet(IEnumerable<T> items, IEqualityComparer<T>? equalityComparer)
         {
-            var multiSet = new MultiSet(equalityComparer);
+            var multiSet = new MultiSet<T>(equalityComparer);
             foreach (T item in items)
             {
                 multiSet.Add(item);
@@ -1395,7 +1395,7 @@ namespace System.Collections.Immutable
 
         private ICollection<int>? FindOrderedIndicesToRemoveByMultiSet(ReadOnlySpan<T> items, IEqualityComparer<T>? equalityComparer)
         {
-            var multiSet = new MultiSet(equalityComparer);
+            var multiSet = new MultiSet<T>(equalityComparer);
             foreach (ref readonly T item in items)
             {
                 multiSet.Add(item);
@@ -1404,7 +1404,7 @@ namespace System.Collections.Immutable
             return GetOrderedIndicesToRemoveFor(multiSet);
         }
 
-        private ICollection<int>? GetOrderedIndicesToRemoveFor(MultiSet multiSet)
+        private ICollection<int>? GetOrderedIndicesToRemoveFor(MultiSet<T> multiSet)
         {
             List<int>? indicesToRemove = null;
             for (int i = 0; i < array!.Length; i++)
@@ -1431,133 +1431,6 @@ namespace System.Collections.Immutable
             }
 
             return indicesToRemove;
-        }
-
-        private class MultiSet
-        {
-            private readonly IEqualityComparer<T> _equalityComparer;
-            private int[] _bucketHeaders; // initialized as zeroed array so its valid item value is 1-based index of _entries
-            private Entry[] _entries;
-            private int _startOffset;
-
-            public MultiSet(int capacity, IEqualityComparer<T>? equalityComparer)
-            {
-                _equalityComparer = equalityComparer ?? EqualityComparer<T>.Default;
-
-                _bucketHeaders = new int[capacity];
-                _entries = new Entry[capacity];
-            }
-
-            public MultiSet(IEqualityComparer<T>? equalityComparer)
-            {
-                _equalityComparer = equalityComparer ?? EqualityComparer<T>.Default;
-
-                _bucketHeaders = new int[1];
-                _entries = new Entry[1];
-            }
-
-            public void Add(T item)
-            {
-                ref int bucketHeader = ref GetBucketHeaderIndexRef(item);
-                int entryIndex = bucketHeader - 1;
-                while (entryIndex != -1)
-                {
-                    ref Entry entry = ref _entries[entryIndex];
-                    if (_equalityComparer.Equals(entry.Key, item))
-                    {
-                        entry.Count++;
-                        return;
-                    }
-
-                    entryIndex = entry.Next;
-                }
-
-                if (_startOffset == _bucketHeaders.Length)
-                {
-                    EnsureCapacity();
-                    bucketHeader = ref GetBucketHeaderIndexRef(item);
-                }
-
-                _entries[_startOffset].SetInfo(item, bucketHeader - 1, 1);
-                bucketHeader = ++_startOffset; // 1-based bucketHeader
-            }
-
-            private void EnsureCapacity()
-            {
-                const uint arrayMaxLength = 0X7FFFFFC7;
-
-                int expectedSize = _bucketHeaders.Length + 1;
-                uint newSize = (uint)_bucketHeaders.Length << 1;
-                if (newSize > arrayMaxLength)
-                {
-                    newSize = arrayMaxLength;
-                }
-
-                if (newSize < expectedSize)
-                {
-                    newSize = (uint)expectedSize;
-                }
-
-                var entries = new Entry[newSize];
-                Array.Copy(_entries, entries, _entries.Length);
-                _bucketHeaders = new int[newSize];
-                for (int i = 0; i < _entries.Length; i++)
-                {
-                    ref Entry entry = ref _entries[i];
-                    if (entry.Count != 0)
-                    {
-                        ref int bucketHeader = ref GetBucketHeaderIndexRef(entry.Key);
-                        entry.Next = bucketHeader - 1;
-                        bucketHeader = i + 1;
-                    }
-                }
-
-                _entries = entries;
-            }
-
-            public bool TryRemove(T item)
-            {
-                ref int bucketHeader = ref GetBucketHeaderIndexRef(item);
-                int entryIndex = bucketHeader - 1;
-                while (entryIndex != -1)
-                {
-                    ref Entry entry = ref _entries[entryIndex];
-                    if (_equalityComparer.Equals(entry.Key, item))
-                    {
-                        if (entry.Count == 0)
-                        {
-                            return false;
-                        }
-
-                        entry.Count--;
-                        return true;
-                    }
-
-                    entryIndex = entry.Next;
-                }
-
-                return false;
-            }
-
-            private ref int GetBucketHeaderIndexRef(T item)
-            {
-                int hashCode = item == null ? 0 : _equalityComparer.GetHashCode(item);
-                return ref _bucketHeaders[(uint)hashCode % _bucketHeaders.Length];
-            }
-
-            private struct Entry
-            {
-                public void SetInfo(T key, int next, int count)
-                {
-                    Key = key;
-                    Next = next;
-                    Count = count;
-                }
-
-                public T Key;
-                public int Next; // 0 based index; -1 means current entry is the last one
-                public int Count;
-            }
         }
     }
 }
